Use context in useLoggedInUser
