+++ src/App.tsx
@@ -3,6 +3,7 @@ import ThemeProvider from '@mui/material/styles/ThemeProvider';
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
 import Navbar from './components/Navbar/Navbar';
+import { UserProvider } from './hooks/useLoggedInUser';
 import Home from './pages/Home';
 import Login from './pages/Login';
 import Sets from './pages/Sets';
@@ -22,23 +23,25 @@ const App = () => {
 	const hiddenPages: AppPage[] = [];
 
 	return (
-		<ThemeProvider theme={theme}>
-			<BrowserRouter>
-				<Navbar title="Flash Cards" pages={navbarPages} />
-				<Container maxWidth="xl" sx={{ mt: 2 }}>
-					<Routes>
-						{[...navbarPages, ...hiddenPages].map(page => (
-							<Route
-								key={page.name}
-								path={page.path}
-								element={page.component}
-							/>
-						))}
-						<Route path="/login" element={<Login />} />
-					</Routes>
-				</Container>
-			</BrowserRouter>
-		</ThemeProvider>
+		<UserProvider>
+			<ThemeProvider theme={theme}>
+				<BrowserRouter>
+					<Navbar title="Flash Cards" pages={navbarPages} />
+					<Container maxWidth="xl" sx={{ mt: 2 }}>
+						<Routes>
+							{[...navbarPages, ...hiddenPages].map(page => (
+								<Route
+									key={page.name}
+									path={page.path}
+									element={page.component}
+								/>
+							))}
+							<Route path="/login" element={<Login />} />
+						</Routes>
+					</Container>
+				</BrowserRouter>
+			</ThemeProvider>
+		</UserProvider>
 	);
 };
